requirements.txt
# Core API & Framework
fastapi>=0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
python-multipart==0.0.12

# AI & LLM
google-generativeai==0.8.3
langchain-core==0.3.12
tiktoken==0.7.0

# UI
gradio==5.7.1

# Text Parsing & OCR
numpy>=1.24.0,<2.0.0
rapidocr_onnxruntime==1.3.24
pdf2image==1.17.0
pdfminer.six==20231228
python-docx==1.1.2
requests==2.32.3
Pillow==10.4.0

# Vector DB & Retrieval
qdrant-client==1.12.1
rank-bm25==0.2.2
